Tyler's First Big Push
So, this is the first iteration of the Main Inventory system, as well as a basic player movement and mouse movement script.

Within the inventory scripts (modular system) it is scalable to handle dragging and dropping between different UI's, and even though there are only two items now (copper and wood) the prefab we can create more from the scriptable object and add them to the enum as we go, and then edit the scriptable object with the necessary info.

Final Note: the added resource folder for fbx'es was a great starting point and we will have to bring the material variable back in when we have more models.

CHECK THE COMMENTS FOR CLARIFICATIONS ON EVERYTHING.