Massive Push
Basically This is a readjustment of the inventory system (no longer adds as needed and shows empty slots the whole time now)

Once a craftable data item (basically a tool item) is in the inventory and selected with a click, you can press 1, 2 or 3 to add it to the hot bar. then out side of the hotbar, you can equip it in your hand.

The make shift craft machine has it's own inventory now which shows you what can be crafted and when the picture is clicked will start crafting ONLY when the generator is running (because of the singleton MachineInventoryManager checking the GeneratorFuelBurner)

All we need to do to add more craftable items is create them with the right click and item the values (and a prefab to spawn with the PickUpItem cause it was being weird)

Nothing is commented as much as I'd like SORRY! I was rushing, I'll fix it tomorrow.