Gamey Updates
Added Main Menu & Pause Menu.

Pause Button is P

Fixed some UI Buggies (hopefully).

Added background music loops and animation that looks like it's swinging the tool as well as sounds to tools.

brought in the logic of the Generator & crafting machine into MainLevel (previously TerrainTest)

Switched the movement to rigidbody velocity logic, added a jump
(left in old logic as comments if we want it back)

added  door manager so we can exit without climbing in the way we came afterwards

Add a bunch of new sprites (minus stones IDK what do with those...)

filled out the craftable data base with all items and added a rock item'

Mapped an UnEquip to Q so you can free your hand whenever (probably need to check if something is in there, but should be fine)

The Rocks in world are straight prefabs placed so should be fine, but I still haven't found  great way around our trees....